feat(VRCStubs): restructure stub folder structure to allow for better organisation
