+++ Runtime/Stubs/VRC_stub/Physbone.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f7cd72a20dbd89f4db4c54271c42b24c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
