+++ Runtime/Stubs/VRC_stub/Physbone/VRCPhysBoneCollider.cs
@@ -1,0 +1,17 @@
+// if for some reason the VRC SDK is in the project then disable this stub
+#if !VRC_SDK_VRCSDK3
+
+using System;
+using System.Collections.Generic;
+using UnityEngine;
+using uk.novavoidhowl.dev.cvrfury.packagecore;
+
+namespace uk.novavoidhowl.dev.vrcstub
+{
+  [AddComponentMenu("")] // hide from add component menu
+  public class VRCPhysBoneCollider : VRCPhysBoneColliderBase
+  {
+  }
+}
+
+#endif
