+++ Runtime/Stubs/VRC_stub/Physbone/VRCPhysBoneCore.cs.meta
