+++ package.json
@@ -22,5 +22,5 @@
   },
   "unity": "2021.3",
   "unityRelease": "23f1",
-  "version": "0.174.1-alpha"
+  "version": "0.175.0-alpha"
 }
